Rework HTTPClient
I'm not really sure ho this module this was working before. The HTTP action functions (e.g.,
get, post, delete) were returning the body of a %Tesla.Env. This was then piped to a
_build_entity function that was invoking body on input (which was already the response body). Because _build_entity functions
use pattern matching to match on a struct, I was seeing match errors as the body of
the response was JSON decoded by the JSON middleware and not a struct. Even if
I implemented _build_entity to match a map (which is what the JSON middleware
was returning), I'd encounter an exception attempting to call Poison.decode! on an
already decoded JSON string.

I am very likely missing something fundamental, but I'm not sure how versions from
0.4.4 to 0.6.0 worked at all. The specs pass because they mock the Tesla request and response and return a Tesla.Env that has another Tesla.Env as its body. I don't believe this
mocks represent the actual responses from the HTTPClient though.

To get things working, I made the following changes:

* Use Tesla.Middleware.EncodeJson instead of Tesla.Middleware.JSON so we can
rely on the middelware to encode requests, but can rely on Poison to decode the response
bodies
* Action functions (delete/delete/post/put) return a Tesla.Env struct
* _build_entity functions match on a Tesla.Env, then use Poison.decode! to decode
the body string as the specified struct
* Updated all of the specs that were mocking Tesla responses to return strings as response
bodies